refactor: remove unused variables
