+++ src/modules/builder/components/common/DuplicateButton.tsx
@@ -3,7 +3,7 @@ import { useTranslation } from 'react-i18next';
 
 import { ListItemIcon, MenuItem } from '@mui/material';
 
-import { DiscriminatedItem } from '@graasp/sdk';
+import { DiscriminatedItem, getParentFromPath } from '@graasp/sdk';
 
 import { CopyIcon } from 'lucide-react';
 
@@ -12,7 +12,6 @@ import { mutations } from '@/config/queryClient';
 import { ITEM_MENU_DUPLICATE_BUTTON_CLASS } from '@/config/selectors';
 
 import { BUILDER } from '~builder/langs';
-import { getParentsIdsFromPath } from '~builder/utils/item';
 
 const DuplicateButton = ({
   item,
@@ -23,11 +22,7 @@ const DuplicateButton = ({
   const { t: translateBuilder } = useTranslation(NS.Builder);
 
   const handleDuplicate = () => {
-    const parentsIds = getParentsIdsFromPath(item.path, { ignoreSelf: true });
-    // get the close parent if not then undefined
-    const to = parentsIds.length
-      ? parentsIds[parentsIds.length - 1]
-      : undefined;
+    const to = getParentFromPath(item.path);
 
     copyItems({
       ids: [item.id],
