+++ src/modules/builder/utils/item.ts
@@ -1,44 +1,3 @@
-// synchronous functions to manage items from redux
-import {
-  DiscriminatedItem,
-  ItemMembership,
-  ItemType,
-  getAppExtra,
-} from '@graasp/sdk';
-
-export const getParentsIdsFromPath = (
-  path?: string,
-  { ignoreSelf = false } = {},
-): string[] => {
-  if (!path) {
-    return [];
-  }
-
-  let p = path;
-  // ignore self item in path
-  if (ignoreSelf) {
-    // split path in half parents / self
-    // eslint-disable-next-line no-useless-escape
-    const els = path.split(/\.[^\.]*$/);
-    // if els has only one element, the item has no parent
-    if (els.length <= 1) {
-      return [];
-    }
-    [p] = els;
-  }
-  const ids = p.replace(/_/g, '-').split('.');
-  return ids;
-};
-
-export const getDirectParentId = (path: string): string | null => {
-  const ids = getParentsIdsFromPath(path);
-  const parentIdx = ids.length - 2;
-  if (parentIdx < 0) {
-    return null;
-  }
-  return ids[parentIdx];
-};
-
 export const LINK_REGEX = new RegExp(
   '^(https?:\\/\\/)?' + // protocol is optional
     '((([a-z\\d]([a-z\\d-]*[a-z\\d])*)\\.)+[a-z]{2,}|' + // domain name
@@ -62,36 +21,6 @@ export const isUrlValid = (str?: string | null): boolean => {
   return false;
 };
 
-/**
- *
- * @deprecated
- */
-export const isItemValid = (item: Partial<DiscriminatedItem>): boolean => {
-  if (!item) {
-    return false;
-  }
-
-  const shouldHaveName = Boolean(item.name?.trim());
-
-  // item should have a type
-  let hasValidTypeProperties =
-    item.type && Object.values<string>(ItemType).includes(item.type);
-  switch (item.type) {
-    case ItemType.APP: {
-      let url;
-      if (item.extra) {
-        ({ url } = getAppExtra(item.extra) || {});
-      }
-      hasValidTypeProperties = isUrlValid(url);
-      break;
-    }
-    default:
-      break;
-  }
-
-  return shouldHaveName && Boolean(hasValidTypeProperties);
-};
-
 export const stripHtml = (str?: string | null): string =>
   str?.replace(/<[^>]*>?/gm, '') || '';
 
@@ -103,36 +32,4 @@ export const sortByName = (
   if (a.name < b.name) return -1;
   if (a.name > b.name) return 1;
   return 0;
-};
-
-// todo: to remove
-// get highest permission a member have over an item,
-// longer the itemPath, deeper is the permission, thus highest
-export const getHighestPermissionForMemberFromMemberships = ({
-  memberships,
-  memberId,
-  itemPath,
-}: {
-  memberships?: ItemMembership[];
-  memberId?: string;
-  itemPath: DiscriminatedItem['path'];
-}): null | ItemMembership => {
-  if (!memberId) {
-    return null;
-  }
-
-  const itemMemberships = memberships?.filter(
-    ({ item: { path: mPath }, account: { id: mId } }) =>
-      mId === memberId && itemPath.includes(mPath),
-  );
-  if (!itemMemberships || itemMemberships.length === 0) {
-    return null;
-  }
-
-  const sorted = [...itemMemberships];
-
-  // sort memberships by the closest to you first (longest path)
-  sorted?.sort((a, b) => (a.item.path.length > b.item.path.length ? -1 : 1));
-
-  return sorted[0];
 };
